+++ db/migrate/20140530133955_add_first_locked_by_id_to_sheets.rb
@@ -1,0 +1,5 @@
+class AddFirstLockedByIdToSheets < ActiveRecord::Migration
+  def change
+    add_column :sheets, :first_locked_by_id, :integer
+  end
+end